photo_microservice/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
Pillow==10.4.0
python-multipart==0.0.9
requests==2.33.1
python-dotenv==1.0.1